# api-gateway/docs/swagger.yaml
definitions:
  flashSale_proto.CreateFlashSaleRequest:
    properties:
      description:
        type: string
      end_time:
        type: string
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/flashSale_proto.ProductId'
        type: array
      start_time:
        type: string
    type: object
  flashSale_proto.CreateFlashSaleResponse:
    properties:
      event:
        $ref: '#/definitions/flashSale_proto.FlashSaleEvent'
    type: object
  flashSale_proto.DeleteFlashSaleRequest:
    properties:
      id:
        type: string
    type: object
  flashSale_proto.DeleteFlashSaleResponse:
    properties:
      message:
        type: string
    type: object
  flashSale_proto.FlashSaleEvent:
    properties:
      description:
        type: string
      end_time:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/flashSale_proto.FlashSaleProduct'
        type: array
      start_time:
        type: string
    type: object
  flashSale_proto.FlashSaleProduct:
    properties:
      description:
        type: string
      discount_price:
        type: number
      is_active:
        type: boolean
      name:
        type: string
      original_price:
        type: number
      product_id:
        type: string
      stock:
        type: integer
    type: object
  flashSale_proto.GetActiveFlashSalesResponse:
    properties:
      active_events:
        items:
          $ref: '#/definitions/flashSale_proto.FlashSaleEvent'
        type: array
    type: object
  flashSale_proto.ListFlashSalesResponse:
    properties:
      events:
        items:
          $ref: '#/definitions/flashSale_proto.FlashSaleEvent'
        type: array
    type: object
  flashSale_proto.ProductId:
    properties:
      descount_price:
        type: number
      product_id:
        type: string
    type: object
  flashSale_proto.PurchaseProductRequest:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
      user_id:
        type: string
    type: object
  flashSale_proto.PurchaseProductResponse:
    properties:
      message:
        type: string
      product:
        $ref: '#/definitions/flashSale_proto.FlashSaleProduct'
    type: object
  flashSale_proto.UpdateFlashSaleRequest:
    properties:
      description:
        type: string
      end_time:
        type: string
      id:
        type: string
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/flashSale_proto.FlashSaleProduct'
        type: array
      start_time:
        type: string
    type: object
  flashSale_proto.UpdateFlashSaleResponse:
    properties:
      event:
        $ref: '#/definitions/flashSale_proto.FlashSaleEvent'
    type: object
  inventory_proto.CreateInventoryReq:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
    type: object
  inventory_proto.CreateInventoryRes:
    properties:
      message:
        type: string
    type: object
  inventory_proto.GetAllInventoriesRes:
    properties:
      inventories:
        items:
          $ref: '#/definitions/inventory_proto.Inventory'
        type: array
    type: object
  inventory_proto.Inventory:
    properties:
      created_at:
        type: string
      inventoryId:
        type: string
      is_active:
        type: boolean
      product_id:
        type: string
      quantity:
        type: integer
      updated_at:
        type: string
    type: object
  inventory_proto.UpdateInventoryRequest:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
    type: object
  inventory_proto.UpdateInventoryResponse:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server for a restaurant reservation system.
  title: Flash-Sale
  version: "1.0"
paths:
  /flashsale:
    delete:
      consumes:
      - application/json
      description: Delete an existing flash sale
      parameters:
      - description: Delete flash sale request body
        in: body
        name: flashsale
        required: true
        schema:
          $ref: '#/definitions/flashSale_proto.DeleteFlashSaleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/flashSale_proto.DeleteFlashSaleResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Delete a flash sale
      tags:
      - flashsale
    get:
      description: Get a list of all flash sales
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/flashSale_proto.ListFlashSalesResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get all flash sales
      tags:
      - flashsale
    post:
      consumes:
      - application/json
      description: Create a new flash sale
      parameters:
      - description: Create flash sale request body
        in: body
        name: flashsale
        required: true
        schema:
          $ref: '#/definitions/flashSale_proto.CreateFlashSaleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/flashSale_proto.CreateFlashSaleResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Create a new flash sale
      tags:
      - flashsale
    put:
      consumes:
      - application/json
      description: Update an existing flash sale
      parameters:
      - description: Update flash sale request body
        in: body
        name: flashsale
        required: true
        schema:
          $ref: '#/definitions/flashSale_proto.UpdateFlashSaleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/flashSale_proto.UpdateFlashSaleResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Update a flash sale
      tags:
      - flashsale
  /flashsale/active:
    get:
      description: Get a list of active flash sales
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/flashSale_proto.GetActiveFlashSalesResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get active flash sales
      tags:
      - flashsale
  /flashsale/purchase:
    post:
      consumes:
      - application/json
      description: Purchase a product that is part of a flash sale
      parameters:
      - description: Purchase product request body
        in: body
        name: purchase
        required: true
        schema:
          $ref: '#/definitions/flashSale_proto.PurchaseProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/flashSale_proto.PurchaseProductResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Purchase a product during a flash sale
      tags:
      - flashsale
  /inventory:
    get:
      description: Get all inventories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inventory_proto.GetAllInventoriesRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get all inventories
      tags:
      - inventory
    post:
      consumes:
      - application/json
      description: Create a new inventory
      parameters:
      - description: inventory request body
        in: body
        name: inventory
        required: true
        schema:
          $ref: '#/definitions/inventory_proto.CreateInventoryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inventory_proto.CreateInventoryRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Create a new inventory
      tags:
      - inventory
    put:
      consumes:
      - application/json
      description: Update an existing inventory
      parameters:
      - description: Update inventory request body
        in: body
        name: inventory
        required: true
        schema:
          $ref: '#/definitions/inventory_proto.UpdateInventoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inventory_proto.UpdateInventoryResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Update an inventory
      tags:
      - inventory
  /inventory/{id}:
    get:
      description: Get an inventory by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inventory_proto.Inventory'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get inventory by ID
      tags:
      - inventory
...
securityDefinitions:
  Bearer:
    description: Enter the token in the format `Bearer {token}`
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"